Extract duplicate code regarding emergency invoice number generation logic

See merge request dev-team/finance-service!33

+++ src/app/Repositories/Invoice/Interface/ItemRepositoryInterface.php
@@ -11,7 +11,7 @@ interface ItemRepositoryInterface extends EloquentRepositoryInterface
 {
     public function indexItemByCriteria(array $criteria, int $limit = 100): Collection;
 
-    public function sumAmountByInvoice(Invoice $invoice): int;
+    public function sumAmountByInvoice(Invoice $invoice): float;
 
     public function indexByInvoices(array $invoiceIds): Collection;
